requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.23.5
streamlit>=1.20.0
matplotlib==3.7.1
tqdm==4.65.0
